third_party/blink/web_tests/webaudio/unit-tests/audit-failures-expected.txt

CONSOLE ERROR: [audit.js] this test requires the explicit comparison with the expected result when it runs with run_web_tests.py.
This is a testharness.js-based test.
[FAIL] X Setting foo to bar incorrectly threw ReferenceError: "bar is not defined".
  assert_true: expected true got false
[FAIL] X function() {\n              let foo = 0;\n            } did not throw an exception.
  assert_true: expected true got false
[FAIL] X 3 < 5 is not true. Got false.
  assert_true: expected true got false
[FAIL] X true is not false. Got true.
  assert_true: expected true got false
[FAIL] X The number of channels is not equal to 2. Got 1.
  assert_true: expected true got false
[FAIL] X 1 should not be equal to 1. Got 1.
  assert_true: expected true got false
[FAIL] X 1 is not NaN but should be. Got 1.
  assert_true: expected true got false
[FAIL] X NaN is NaN but should not be. Got NaN.
  assert_true: expected true got false
[FAIL] X typeof 3.141592 is not equal to string. Got number.
  assert_true: expected true got false
[FAIL] X 1 is not greater than 2. Got 1.
  assert_true: expected true got false
[FAIL] X 1 is not greater than or equal to 2. Got 1.
  assert_true: expected true got false
[FAIL] X 2 is not less than 1. Got 2.
  assert_true: expected true got false
[FAIL] X 2 is not less than or equal to 1. Got 2.
  assert_true: expected true got false
[FAIL] X 1 is not equal to 2. Got 1.
  assert_true: expected true got false
[FAIL] X should(1).beEqualTo(2) is not true. Got false.
  assert_true: expected true got false
[FAIL] X The message is false! Got false.
  assert_true: expected true got false
[FAIL] X Decoding audio data with no argument rejected incorrectly with TypeError: Failed to execute 'decodeAudioData' on 'BaseAudioContext': 1 argument required, but only 0 present.. Got Promise.
  assert_true: expected true got false
[FAIL] X Suspending OAC with no argument rejected correctly but got TypeError instead of IndexSizeError. Got Promise.
  assert_true: expected true got false
[FAIL] X Start OAC rendering resolved incorrectly. Got Promise.
  assert_true: expected true got false
[FAIL] < [basic-failure] 19 out of 19 assertions were failed.
  assert_true: expected true got false
[FAIL] X 0 is not close to 0.1 within a relative error of 0 (RelErr=1). Got 0.
  assert_true: expected true got false
[FAIL] X The measured decibel is not close to 62 within a relative error of 0.01 (RelErr=0.04838709677419355). Got 59.
  assert_true: expected true got false
[FAIL] X Test array: Expected 5 for all values but found 7 unexpected values: \n	Index	Actual\n	[0]	1\n	[1]	8\n	[2]	11\n	[3]	18\n	...and 3 more errors.
  assert_true: expected true got false
[FAIL] X [0,0,0] should have contain at least one value different from 0.
  assert_true: expected true got false
[FAIL] X [0,1,2,3,4,5,6,7,8,9] expected to be equal to the array [0,0,0,0,0,0,0,0,0,0] but differs in 9 places:\n	Index	Actual			Expected\n	[1]	1.0000000000000000e+0	0.0000000000000000e+0\n	[2]	2.0000000000000000e+0	0.0000000000000000e+0\n	[3]	3.0000000000000000e+0	0.0000000000000000e+0\n	[4]	4.0000000000000000e+0	0.0000000000000000e+0\n	...and 5 more errors.
  assert_true: expected true got false
[FAIL] X [1,1,1,1,2,2,3,3,3] expected to have the value sequence of [1,3,2] but got 2 at index 4.
  assert_true: expected true got false
[FAIL] X [1,1,2,2] expected to have the value sequence of [1,2,3] but got 2 values, [1,2], instead of 3.
  assert_true: expected true got false
[FAIL] X [1,2,3,4,5] expected to have the value sequence of [1,2,3] but got 5 values, [1,2,3,4,5], instead of 3.
  assert_true: expected true got false
[FAIL] X [0.5,0.5,0.55,0.5,0.45,0.5] has a glitch at index 5 of size 0.04999999999999999.
  assert_true: expected true got false
[FAIL] X [0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.9] does not equal [0.7,0.6,0.5,0.4,0.3,0.2,0.1] with an element-wise tolerance of {"absoluteThreshold":0.02,"relativeThreshold":0}.\n	Index	Actual			Expected		AbsError		RelError		Test threshold\n	[0]	1.0000000000000001e-1	6.9999999999999996e-1	5.9999999999999998e-1	8.5714285714285721e-1	2.0000000000000000e-2\n	[1]	2.0000000000000001e-1	5.9999999999999998e-1	3.9999999999999997e-1	6.6666666666666663e-1	2.0000000000000000e-2\n	[2]	2.9999999999999999e-1	5.0000000000000000e-1	2.0000000000000001e-1	4.0000000000000002e-1	2.0000000000000000e-2\n	[4]	5.0000000000000000e-1	2.9999999999999999e-1	2.0000000000000001e-1	6.6666666666666674e-1	2.0000000000000000e-2\n	[5]	5.9999999999999998e-1	2.0000000000000001e-1	3.9999999999999997e-1	1.9999999999999998e+0	2.0000000000000000e-2\n	...and 1 more errors.\n	Max AbsError of 5.9999999999999998e-1 at index of 0.\n	Max RelError of 5.9999999999999991e+0 at index of 6.\n	[6]	6.9999999999999996e-1	1.0000000000000001e-1	5.9999999999999998e-1	5.9999999999999991e+0	2.0000000000000000e-2\n
  assert_true: expected true got false
[FAIL] X [0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.9] does not equal [0.7,0.6,0.5,0.4,0.3,0.2,0.1] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0.1}.\n	Index	Actual			Expected		AbsError		RelError		Test threshold\n	[0]	1.0000000000000001e-1	6.9999999999999996e-1	5.9999999999999998e-1	8.5714285714285721e-1	6.9999999999999993e-2\n	[1]	2.0000000000000001e-1	5.9999999999999998e-1	3.9999999999999997e-1	6.6666666666666663e-1	5.9999999999999998e-2\n	[2]	2.9999999999999999e-1	5.0000000000000000e-1	2.0000000000000001e-1	4.0000000000000002e-1	5.0000000000000003e-2\n	[4]	5.0000000000000000e-1	2.9999999999999999e-1	2.0000000000000001e-1	6.6666666666666674e-1	2.9999999999999999e-2\n	[5]	5.9999999999999998e-1	2.0000000000000001e-1	3.9999999999999997e-1	1.9999999999999998e+0	2.0000000000000004e-2\n	...and 1 more errors.\n	Max AbsError of 5.9999999999999998e-1 at index of 0.\n	Max RelError of 5.9999999999999991e+0 at index of 6.\n	[6]	6.9999999999999996e-1	1.0000000000000001e-1	5.9999999999999998e-1	5.9999999999999991e+0	1.0000000000000002e-2\n
  assert_true: expected true got false
[FAIL] X [0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.9] does not equal [0.7,0.6,0.5,0.4,0.3,0.2,0.1] with an element-wise tolerance of {"absoluteThreshold":0.02,"relativeThreshold":0.1}.\n	Index	Actual			Expected		AbsError		RelError		Test threshold\n	[0]	1.0000000000000001e-1	6.9999999999999996e-1	5.9999999999999998e-1	8.5714285714285721e-1	6.9999999999999993e-2\n	[1]	2.0000000000000001e-1	5.9999999999999998e-1	3.9999999999999997e-1	6.6666666666666663e-1	5.9999999999999998e-2\n	[2]	2.9999999999999999e-1	5.0000000000000000e-1	2.0000000000000001e-1	4.0000000000000002e-1	5.0000000000000003e-2\n	[4]	5.0000000000000000e-1	2.9999999999999999e-1	2.0000000000000001e-1	6.6666666666666674e-1	2.9999999999999999e-2\n	[5]	5.9999999999999998e-1	2.0000000000000001e-1	3.9999999999999997e-1	1.9999999999999998e+0	2.0000000000000004e-2\n	...and 1 more errors.\n	Max AbsError of 5.9999999999999998e-1 at index of 0.\n	Max RelError of 5.9999999999999991e+0 at index of 6.\n	[6]	6.9999999999999996e-1	1.0000000000000001e-1	5.9999999999999998e-1	5.9999999999999991e+0	2.0000000000000000e-2\n
  assert_true: expected true got false
[FAIL] X [1,2,3,4,5,6,7,8,9,10,11] does not equal [2,3,4,5,6,7,800,900,1000,1100,1200] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0}.\n	Index	Actual			Expected		AbsError		RelError		Test threshold\n	[0]	1.0000000000000000e+0	2.0000000000000000e+0	1.0000000000000000e+0	5.0000000000000000e-1	0.0000000000000000e+0\n	[1]	2.0000000000000000e+0	3.0000000000000000e+0	1.0000000000000000e+0	3.3333333333333331e-1	0.0000000000000000e+0\n	[2]	3.0000000000000000e+0	4.0000000000000000e+0	1.0000000000000000e+0	2.5000000000000000e-1	0.0000000000000000e+0\n	[3]	4.0000000000000000e+0	5.0000000000000000e+0	1.0000000000000000e+0	2.0000000000000001e-1	0.0000000000000000e+0\n	[4]	5.0000000000000000e+0	6.0000000000000000e+0	1.0000000000000000e+0	1.6666666666666666e-1	0.0000000000000000e+0\n	...and 6 more errors.\n	Max AbsError of 1.1890000000000000e+3 at index of 10.\n	[10]	1.1000000000000000e+1	1.2000000000000000e+3	1.1890000000000000e+3	9.9083333333333334e-1	0.0000000000000000e+0\n	Max RelError of 9.9124999999999996e-1 at index of 6.\n	[6]	7.0000000000000000e+0	8.0000000000000000e+2	7.9300000000000000e+2	9.9124999999999996e-1	0.0000000000000000e+0\n
  assert_true: expected true got false
[FAIL] < [numerical-failures] 13 out of 13 assertions were failed.
  assert_true: expected true got false
[FAIL] X Loading non-existent file within should().beResolved() rejected incorrectly with loadFile: Network failure when loading non-existent-audio-file.wav.. Got Promise.
  assert_true: expected true got false
[FAIL] < [load-file-in-should] 1 out of 1 assertions were failed.
  assert_true: expected true got false
[FAIL] # AUDIT TASK RUNNER FINISHED: 3 out of 3 tasks were failed.
  assert_true: expected true got false
Harness: the test ran to completion.
